+++ Plugin/src/Patches/PlayerControllerBPatch.cs
@@ -1,6 +1,6 @@
-using CodeRebirth.src.Content.Enemies;
+using System.Collections.Generic;
 using CodeRebirth.src.Content.Items;
-using CodeRebirth.src.Content.Unlockables;
+using CodeRebirth.src.MiscScripts;
 using CodeRebirth.src.Util;
 using GameNetcodeStuff;
 using HarmonyLib;
@@ -11,6 +11,8 @@ using UnityEngine;
 namespace CodeRebirth.src.Patches;
 [HarmonyPatch(typeof(PlayerControllerB))]
 static class PlayerControllerBPatch {
+    public static List<SmartAgentNavigator> smartAgentNavigators = new();
+
     [HarmonyPatch(nameof(PlayerControllerB.PlayerHitGroundEffects)), HarmonyPrefix]
     public static void PlayerHitGroundEffects(PlayerControllerB __instance) {
         if (!__instance.ContainsCRPlayerData()) return;
@@ -50,21 +52,10 @@ static class PlayerControllerBPatch {
 
     private static void PlayerControllerB_TeleportPlayer(On.GameNetcodeStuff.PlayerControllerB.orig_TeleportPlayer orig, PlayerControllerB self, Vector3 pos, bool withRotation, float rot, bool allowInteractTrigger, bool enableController)
     {
-        foreach (var enemy in RoundManager.Instance.SpawnedEnemies)
+        foreach (var navigator in smartAgentNavigators)
         {
-            if (enemy is CodeRebirthEnemyAI codeRebirthEnemyAI)
-            {
-                Plugin.ExtendedLogging($"Setting codeRebirthEnemyAI.positionsOfPlayersBeforeTeleport[self] to {self.transform.position}");
-                codeRebirthEnemyAI.positionsOfPlayersBeforeTeleport[self] = self.transform.position;
-            }
-        }
-        foreach (ShockwaveGalAI gal in UnityEngine.Object.FindObjectsOfType<ShockwaveGalAI>())
-        {
-            if (self == gal.ownerPlayer)
-            {
-                Plugin.ExtendedLogging($"Setting gal.positionOfPlayerBeforeTeleport to {self.transform.position}");
-                gal.positionOfPlayerBeforeTeleport = self.transform.position;
-            }
+            Plugin.ExtendedLogging($"Setting SmartAgentNavigator.positionsOfPlayersBeforeTeleport[self] to {self.transform.position}");
+            navigator.positionsOfPlayersBeforeTeleport[self] = self.transform.position;
         }
         orig(self, pos, withRotation, rot, allowInteractTrigger, enableController);
     }